requirements.txt
pygame-ce==2.5.7
pyinstaller==6.19.0
typing_extensions==4.15.0
